Worked on Board Details, Player hitboxes.
Altered the collider for the Players, as balls would bounce off of
nothing, which is visually confusing.

Messed around with the position of the Score canvas on the board, and
added a black background to it, to make it look like a part of the
board.

Also disabled from Debug logging, as there were hundreds of debug
messages.