Main menu, GameManager, and Chomp Effect changes
The Chomp effects now render on top of everything else in the scene.
This was done by having an extra camera at a higher depth value
rendering only the Chomp Effects.

Made some changes to the game manager to better handle scene changes. It
also forces the game to be in Portrait perspective only! Also added code
to prevent some sounds from playing on the Main Menu.

Made the main menu look nicer by added a title, a background, and some
details to it. There may be some wonkyness with the subtitle, but it was
fine on my phone.