Fixed the Win Screen and made it nicer
There were some problems with reaching the win screen. Players can now
loop through the game properly from the main menu to the win screen.
There was also a problem with getting the win screen to say the right
thing, but i managed to get it working.

Also added a colored background, which will gain the color of the
victor's side of the board.